MAINT: rename pytorch layer.py -> operator.py
