+++ odl/contrib/pytorch/examples/operator.py
